decouple context-aware functions
